fix: claim

+++ services/admin-api/src/mechanics/claim/claim.service.ts
@@ -1,30 +1,23 @@
 import { Inject, Injectable, Logger, LoggerService, NotFoundException } from "@nestjs/common";
-import { ConfigService } from "@nestjs/config";
 import { InjectRepository } from "@nestjs/typeorm";
 import { DeleteResult, FindOneOptions, FindOptionsWhere, Repository } from "typeorm";
-import { utils, Wallet } from "ethers";
+import { utils } from "ethers";
 
 import { ClaimStatus, IClaimSearchDto, TokenType } from "@framework/types";
-import { ETHERS_SIGNER } from "@gemunion/nestjs-ethers";
+import { SignerService } from "@gemunion/nest-js-module-exchange-signer";
 
 import { IClaimItemCreateDto } from "./interfaces";
 import { ClaimEntity } from "./claim.entity";
 import { AssetService } from "../asset/asset.service";
-import { TemplateService } from "../../blockchain/hierarchy/template/template.service";
-import { SignerService } from "../signer/signer.service";
 
 @Injectable()
 export class ClaimService {
   constructor(
-    @Inject(ETHERS_SIGNER)
-    private readonly signer: Wallet,
     @Inject(Logger)
     private readonly loggerService: LoggerService,
-    private readonly configService: ConfigService,
     @InjectRepository(ClaimEntity)
     private readonly claimEntityRepository: Repository<ClaimEntity>,
     protected readonly assetService: AssetService,
-    protected readonly templateService: TemplateService,
     private readonly signerService: SignerService,
   ) {}
 
@@ -85,12 +78,12 @@ export class ClaimService {
   public async update(where: FindOptionsWhere<ClaimEntity>, dto: IClaimItemCreateDto): Promise<ClaimEntity> {
     const { account, item } = dto;
 
-    const claimEntity = await this.findOne(where, {
+    let claimEntity = await this.findOne(where, {
       join: {
         alias: "claim",
         leftJoinAndSelect: {
           item: "claim.item",
-          components: "item.components",
+          item_components: "item.components",
         },
       },
     });
@@ -106,12 +99,20 @@ export class ClaimService {
 
     await this.assetService.update(claimEntity.item, item);
 
-    const templateEntity = await this.templateService.findOne(
-      { id: item.components[0].templateId },
-      { relations: { contract: true } },
-    );
-    if (!templateEntity) {
-      throw new NotFoundException("templateNotFound");
+    claimEntity = await this.findOne(where, {
+      join: {
+        alias: "claim",
+        leftJoinAndSelect: {
+          item: "claim.item",
+          item_components: "item.components",
+          item_template: "item_components.template",
+          item_contract: "item_components.contract",
+        },
+      },
+    });
+
+    if (!claimEntity) {
+      throw new NotFoundException("claimNotFound");
     }
 
     const nonce = utils.randomBytes(32);
@@ -124,6 +125,16 @@ export class ClaimService {
   }
 
   public async delete(where: FindOptionsWhere<ClaimEntity>): Promise<DeleteResult> {
+    const claimEntity = await this.findOne(where);
+
+    if (!claimEntity) {
+      throw new NotFoundException("claimNotFound");
+    }
+
+    if (claimEntity.claimStatus !== ClaimStatus.NEW) {
+      throw new NotFoundException("claimRedeemed");
+    }
+
     return this.claimEntityRepository.delete(where);
   }
 
@@ -141,7 +152,7 @@ export class ClaimService {
       claimEntity.item.components.map(component => ({
         tokenType: Object.keys(TokenType).indexOf(component.tokenType),
         token: component.contract.address,
-        tokenId: component.template.tokens[0].tokenId,
+        tokenId: component.template.id.toString(),
         amount: component.amount,
       })),
       [],
